refactor: Business Profile 파일구조 수정 (#239)
