--- frontend/src/components/BProfile/BOtherProfile/index.jsx
@@ -1,10 +1,0 @@
-import React from "react";
-import PBOtherProfile from "./PBOtherProfile";
-
-export default function BOtherProfile({ businessId }) {
-  return (
-    <>
-      <PBOtherProfile originalId={businessId} />
-    </>
-  );
-}
