+++ frontend/src/components/BProfile/BProfileInfo/BProfileEditModal/presenter.jsx
@@ -1,67 +1,51 @@
-import React, { useState } from "react";
-import Portal from "../../Common/Modal/Portal";
-import { PVImg } from "../../Common";
-import PropTypes from "prop-types";
+import React from "react";
+import Portal from "../../../Common/Modal/Portal";
 import CloseIcon from "@mui/icons-material/Close";
-import * as S from "../style";
-import Address from "../../Common/Address";
-import { handleCompressImg } from "../../../utils/imgCompress";
-import { useRecoilValue } from "recoil";
-import { userState } from "../../../recoil/userState";
 
-BCreateCont.propTypes = {
-  setModalSwitch: PropTypes.func.isRequired,
-  CreateBProfile: PropTypes.func.isRequired,
-};
+import { PVImg } from "../../../Common";
+import Address from "../../../Common/Address";
 
-export default function BCreateCont({ setModalSwitch, CreateBProfile }) {
-  const [inputs, setInputs] = useState({
-    name: "",
-    place: "",
-    introduce: "afewf",
-    contact: "",
-  });
-  const [img, setImg] = useState();
-  const [admodalSwitch, setadModalSwitch] = useState(false);
-  const admin = useRecoilValue(userState).name;
-  const handleAddress = async (fullAddress) => {
-    setInputs({
-      ...inputs,
-      place: fullAddress,
-    });
-  };
-  const { name, place, contact } = inputs;
-  const onChange = (e) => {
-    const { value, name } = e.target;
-    setInputs({
-      ...inputs,
-      [name]: value,
-    });
-  };
-  const closeModal = () => {
-    setModalSwitch(false);
-  };
-  const fileChange = (e) => {
-    handleCompressImg(e.target.files[0]).then((res) => setImg(res));
-  };
-  const imgDel = () => {
-    setImg(undefined);
-  };
+import * as S from "../../style";
+
+export default function BProfileEditCont({
+  closeModal,
+  basicImg,
+  newImg,
+  BProfileData,
+  imgDel,
+  fileChange,
+  name,
+  onChange,
+  admin,
+  place,
+  setadModalSwitch,
+  admodalSwitch,
+  handleAddress,
+  contact,
+  bpEditHandler,
+  inputs,
+}) {
   return (
     <Portal>
       <S.ModalBackground>
         <S.ModalBox>
           <S.ModalHeader>
-            <p>비즈니스 프로필 생성하기</p>
+            <p>비즈니스 프로필 수정하기</p>
             <S.CloseModalBtn onClick={closeModal}>
               <CloseIcon />
             </S.CloseModalBtn>
           </S.ModalHeader>
-          <S.ModalContent>
+          <S.ModalContentEdit>
             <S.BNCont1>
               <S.ImgBox>
-                {img ? (
-                  <PVImg img={img} />
+                {basicImg === false ? (
+                  newImg ? (
+                    <PVImg img={newImg} />
+                  ) : BProfileData.avatar ? (
+                    <S.Img src={BProfileData.avatar} />
+                  ) : (
+                    <S.Img src="/기본비즈니스프로필.jpeg" />
+                  )
                 ) : (
                   <S.Img src="/기본비즈니스프로필.jpeg" />
                 )}
@@ -69,8 +53,8 @@ export default function BCreateCont({ setModalSwitch, CreateBProfile }) {
               <S.ImgDel onClick={imgDel}>기본이미지로 변경</S.ImgDel>
               <S.FileLabel htmlFor="input-file">사진 업로드 하기</S.FileLabel>
               <S.BPEditFileInput
-                id="input-file"
                 name="img"
+                id="input-file"
                 type="file"
                 onChange={fileChange}
               />
@@ -129,16 +113,17 @@ export default function BCreateCont({ setModalSwitch, CreateBProfile }) {
                   />
                 </S.BNLabel>
               </S.BNCont23>
+              <S.BPEditButton
+                type="submit"
+                onClick={async (e) => {
+                  e.preventDefault();
+                  bpEditHandler(inputs, newImg, basicImg);
+                }}
+              >
+                <p>수정하기</p>
+              </S.BPEditButton>
             </S.BNCont2>
-          </S.ModalContent>
-          <S.BPEditButton
-            onClick={async (e) => {
-              e.preventDefault();
-              CreateBProfile(inputs, img);
-            }}
-          >
-            <p>생성하기</p>
-          </S.BPEditButton>
+          </S.ModalContentEdit>
         </S.ModalBox>
       </S.ModalBackground>
     </Portal>
