--- frontend/src/components/BProfile/BProfileSection/index.jsx
@@ -1,6 +1,0 @@
-import React from "react";
-import PBProfileSection from "./PBProfileSection";
-
-export default function BProfileSection() {
-  return <PBProfileSection />;
-}
