--- frontend/src/components/BProfileEdit/index.jsx
@@ -1,2 +1,0 @@
-export { default as BProfileEditCont } from "./BProfileEditCont";
-export { default as BCreateCont } from "./BCreateCont";
